\name{Atar}
\alias{Atar}
\docType{data}
\title{Total Backward and Forward Output Linkage Indices}
\description{The input-output analysis is used to examine the significance and impact of tourism sector on theeconomy of Turkey and to identify whether its interaction with other sectors may enhance economic growth.}
\usage{data(Atar)}
\format{
A data frame with 16 observations on the following two variables. Row names indicate the type of economic sector.
  \describe{
  	\item{TBL}{Total Backward Linkage Index.}
  	\item{TFL}{Total Forward Linkage Index}
  	}
  }
  
\details{The total amount of the input from the other sectors for the output in tourism sector is high and tourism sector is in a structure of nurturing other sectors hence, it is thought that tourism sector will support the production in other sectors, with a considerable impact on growth.}

\examples{
	data(Atar)
	head(Atar)
}

\source{
Atan, Y. and Arslanturk, Y. (2012). Tourism and economic growth nexus: an input-output analysis in Turkey. \emph{Procedia - Social and Behavioral Sciences}. \bold{62}, 952-956.
}

\keyword{datasets}
