\name{DietFish}
\alias{DietFish}
\docType{data}
\title{Comparison of total diets of banded killifish and mummichog}
\description{Data collected in Porters Lake (Nova Scotia, Can), from june through august 1971.}
\usage{data(DietFish)}
\format{
A data frame with 31 observations on percentage frequency of a food item in the diet of two species of California killfish: Fundulus diaphanus (banded killifish) and Fundulus heteroclitus (mummichog). Row names indicate the food item.
  \describe{
  	\item{occ_D}{frequency for F. Diaphanus.}
  	\item{occ_H}{frequency for F. Heteroclitus.}
  	}
}  
\details{
Banded killifish ranged in size from 27-107 mm total length and mummichog ranged in size from 35-92 mm total length.	
}

\examples{
	data(DietFish)
	head(DietFish)
}

\source{
Fritz, E. S. (1974). "Total diet comparison in fishes by Spearman rank correlation coefficients".  \emph{Copeia}, \bold{1}, 210--214.}

\keyword{datasets}
