\name{Femurs}
\alias{Femurs}
\docType{data}
\title{Radiographic-caudolateral curvilinear osteophyte}
\description{Caudolateral curvilinear osteophyte (CCO), an osteophyte at the site of joint capsule attachment on the caudal aspect of the femoral neck, has been advocated as a radiographic criterion for coxofemoral subluxation.}
\usage{data(Femurs)}
\format{

  A data frame on the following variables concerning with 44 femurs .
  \describe{
  	\item{CCO}{Caudolateral curvilinear osteophyte index (in \eqn{mm^2}).}
  	\item{Age}{Age (month).}
  	\item{BW}{Body weight (kg).}
  	\item{Gender}{Gender, 1=Female, 2=Male.}  	
  	}}
  	
\examples{
	data(Femurs)
	head(Femurs)
}

\source{
Kishimoto, M. et al. (2010). "The relationship between the size of caudolateral curvilinear osteophyte of the canine femoral neck and the radiographic view". \emph{Journal of Veterinary Science}, \bold{11}, 89--91.
}

\keyword{datasets}
