\name{Gabbs}
\alias{Gabbs}
\docType{data}
\title{Dimensions of fetal metacurpals and phalanges}
\description{Micrometric measurements of optically-projected histological hand sections.}
\usage{data(Gabbs)}
\format{
A data frame with 19 observations on the following 6 variables.
  \describe{
  	\item{V15-29mm}{Absolute measurements in millimeters. }
  	\item{V30-44mm}{Absolute measurements in millimeters. }
  	\item{V45-59mm}{Absolute measurements in millimeters. }
  	\item{V60-74mm}{Absolute measurements in millimeters. }
  	\item{V75-89mm}{Absolute measurements in millimeters. }
  	\item{V90-104mm}{Absolute measurements in millimeters.}
  	\item{AR}{Adult ranks.}
  	}
  }
  
\details{Bone-to-bone ratios or proportions derived from these measurements evidence a very early disproportionate elongation of distals followed by a gradual approximation of adult proportions by the 90-104 mm crown-rump length. Adult ranks calculated from Garn et al. (1972).}

\examples{
	data(Gabbs)
	head(Gabbs)
}

\source{
Garn, S. M. et al. (1975). "Early prenatal rankings and attainment of adult metacarpal-phalangeal proportions". \emph{American Journal of Physical Anthropology}, \bold{43}, 327--332.

Garn, S. M. et al. (1972). "Metacarpophalangeal length in the evaluation of skeletal malformation".  \emph{Radiology}, \bold{105}, 375--381.
}

\keyword{datasets}
