\name{Laudaher}
\alias{Laudaher}
\docType{data}
\title{Data on the significance of lymphocytic infiltration}
\description{Non-parametric rank order statistical methods were used to establish quantitative correlations, particularly withthe duration of survival.}
\usage{data(Laudaher)}
\format{
  A data frame with 16 observations on the following two variables.
  \describe{
  	\item{Duration}{Duration of survival (months).}
  	\item{Infiltration}{Lymphocytic infiltration of tumour (score).}
  	}}
\details{The study shows a significant relationship between lymphocytic infiltration of the tumour and survival of the patient.}
\examples{
	data(Laudaher)
	head(Laudaher)
}

\source{
Lauder, I. and Aherne, A. (1972). "The significance of limphocytic infiltration in neuroblastoma". \emph{British Journal of Cancer}, \bold{26}, 321--330.
}

\keyword{datasets}
