\name{Locre}
\alias{Locre}
\docType{data}
\title{Adult sex differences in crying behavior.}
\description{Female subjects reported crying significantly more frequently and intensely than did male subfects. Despite these sex differences, many evidences of similarity between the sexes in crying behavior were found.}
\usage{data(Locre)}
\format{
Subjects were asked to check any of 20 different situations (selected on the basis of preliminary study) which would be likely to make them cry.
  \describe{
  	\item{Males}{Percentages of subjects.}
  	\item{Females}{Percentages of subjects.}
  	}
  }
  
\details{Subjects were students in introductory psychology and sociology classes, a general education requirement for a variety of majors and schools at California. The mean age of subjects was 22 years.}

\note{Some potential outliers are present, e.g. 1= Death_of_someone_close}

\examples{
	data(Locre)
	head(Locre)
}

\source{
Lombardo, W. K.  and Cretser, G.A. and Lombardo, B. (1983). Fer cryin'out loud - there is a sex difference. \emph{Sex Roles}, \bold{9}, 987-995.
}

\keyword{datasets}
