\name{Marozzi}
\alias{Marozzi}
\docType{data}
\title{Trust in public institutions}
\description{Improving trust in public institutions is central to improve social capital, participation in civic activities and law-abiding behavior and then governmental legitimacy.}
\usage{data(Marozzi)}
\format{
The data are the proportions of citizens of 29 European countries having a positive trust in various public institutions. 
  \describe{
  	\item{Ind_1}{Parliament.}
  	\item{Ind_2}{Legal system.}
  	\item{Ind_3}{Police.}
  	\item{Ind_4}{Politicians.}
  	\item{Ind_5}{Political parties.}
  	\item{Ind_6}{European Parliament.}
  	\item{Ind_7}{United Nations.}
  	}
  }
\details{It is found that Scandinavian countries are the most trustful European countries in public institutions, whereas former communist countries as well as Iberian and Mediterranean ones are much less trustful.}

\examples{
	data(Marozzi)
	head(Marozzi)
}

\source{
Marozzi, M. (2015). "Measuring trust in European public Institutions". \emph{Social Indicators Research}, \bold{123}, 879--895.
}

\keyword{datasets}
