\name{Starshi}
\alias{Starshi}
\docType{data}
\title{Policy-related determinants of health}
\description{International comparisons can provide clues to understanding some of the important policy-related determinants of health including those related to the provision of health care services.}
\usage{data(Starshi)}
\format{
A data frame with 13 observations on the following two variables.
  \describe{
  	\item{Sm15F}{Percentage of females who smoke at ages 15 and older.}
  	\item{Sm15M}{Percentage of males who smoke at ages 15 and older.}
 All countries 1992, except Canada (1991), Spain (1993).
  	}
  }
  
\details{Data on major health indicators were obtained from OECD and World Health Organization published data.}

\examples{
	data(Starshi)
	head(Starshi)
}

\source{
Starfield, B. and Shi, L. (2002). "Policy relevant determinants of health: an international perspective". \emph{Health Policy}, \bold{60}, 201--218.
}

\keyword{datasets}
